requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
blessings==1.7
boto3==1.12.13
botocore==1.15.13
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.4
django-bmemcached==0.3.0
django-celery-beat==2.0.0
django-configurations==2.2
django-cors-headers==3.2.1
django-countries==6.0
django-filter==2.2.0
django-model-utils==4.0.0
django-nose==1.4.6
django-recurrence==1.10.2
django-storages==1.9.1
django-timezone-field==4.0
django-unique-upload==0.2.1
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.1
flake8==3.7.9
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
jmespath==0.9.5
kombu==4.6.8
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.2.1
mccabe==0.6.1
mock==4.0.1
newrelic==5.8.0.136
nose==1.3.7
nose-progressive==1.5.2
oauthlib==3.1.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pylibmc==1.6.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-binary-memcached==0.29.0
python-crontab==2.4.0
python-dateutil==2.8.1
python-twitter==3.5
pytz==2019.3
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
responses==0.10.12
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
typed-ast==1.4.1
uhashring==1.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
wrapt==1.11.2
zipp==3.1.0
